manpage: update MAINTAINERS section
removes a couple maintainers who have never been active in the
development.

there are a couple other maintainers have not been active in a while,
but was somewhat active in the past. so keep their names in it
(for now at least).

also re-arrange the entires a bit based on activity.