document handling of empty X resources values
since fb639bb we reject empty X resources value to support use-cases
like this: https://codeberg.org/nsxiv/nsxiv/issues/339
this issue was also bought up by GRFreire in:
https://codeberg.org/nsxiv/nsxiv-record/pulls/115#issuecomment-474831

this suggests that this is a use-case that a decent amount of users
might be interested in. so document the behavior so it's more easily
visible.