Fix exiv2 argument order in key-handler example (#518)
The exif viewer command (`C-x C-e`) in the `key-handler` example seems to be
wrong. It gives me this output:

	$ exiv2 pr -q -pa SeaSunset.jpg
	-q: Failed to open the file
	-pa: Failed to open the file
	SeaSunset.jpg        File name       : SeaSunset.jpg
	SeaSunset.jpg        File size       : 2247224 Bytes
	...

`-q` and `-pa` are interpreted as filenames. The correct argument order is to
put the _options_ before the _action_:

	$ exiv2 -h
	Usage: exiv2 [ option [ arg ] ]+ [ action ] file ...

With the new order as changed in this pull request, I get this output:

	$ exiv2 -q -pa pr SeaSunset.jpg
	Exif.Image.Make                              Ascii       6  Canon
	Exif.Image.Model                             Ascii      22  Canon EOS 5D Mark III
	...

Tested with exiv2 `0.28.3`.

Reviewed-on: https://codeberg.org/nsxiv/nsxiv/pulls/518
Reviewed-by: eylles <eylles@noreply.codeberg.org>
Co-authored-by: Max Schillinger <maxschillinger@web.de>
Co-committed-by: Max Schillinger <maxschillinger@web.de>